+++ code/src/main/java/io/almostrealism/expression/Epsilon.java
@@ -20,6 +20,7 @@ import io.almostrealism.kernel.KernelStructureContext;
 import io.almostrealism.lang.LanguageOperations;
 
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 public class Epsilon extends StaticReference<Double> {
 
@@ -33,8 +34,8 @@ public class Epsilon extends StaticReference<Double> {
 	}
 
 	@Override
-	public OptionalInt upperBound(KernelStructureContext context) {
-		return OptionalInt.of(1);
+	public OptionalLong upperBound(KernelStructureContext context) {
+		return OptionalLong.of(1);
 	}
 
 	@Override
