+++ code/src/main/java/io/almostrealism/expression/Floor.java
@@ -22,6 +22,7 @@ import io.almostrealism.lang.LanguageOperations;
 import java.util.List;
 import java.util.OptionalDouble;
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 public class Floor extends Expression<Double> {
 	public Floor(Expression<Double> input) {
@@ -45,10 +46,10 @@ public class Floor extends Expression<Double> {
 	}
 
 	@Override
-	public OptionalInt upperBound(KernelStructureContext context) {
-		OptionalInt v = getChildren().get(0).upperBound(context);
+	public OptionalLong upperBound(KernelStructureContext context) {
+		OptionalLong v = getChildren().get(0).upperBound(context);
 		if (v.isPresent()) return v;
-		return OptionalInt.empty();
+		return OptionalLong.empty();
 	}
 
 	@Override
